Refactored and added basic button switch to scoreboard
